Subclass JobTemplateV2 for #launch
https://bugzilla.redhat.com/show_bug.cgi?id=1369842
